requirements.txt
fastapi==0.115.0 
uvicorn==0.31.0 
httpx==0.27.2 
python-dotenv==1.0.1 
pydantic==2.9.2 
pywa==0.0.27 
redis==5.1.0 # Opcional, para gerenciar sessões